wordsquare: Simplistic approach with dynamic wordlist and recursion
This approach tries words in rows, only failing when no words fit
at any of the positions. It does not use that the list is sorted.
The largest problem is probably that it it will try different
words, even when some of the positions can only accept one or two
characters. This should led to serious optimizations.

Obviously the code is pretty scrappy right now, lots of messy
string formatting.